Revert "Add persistent storage for the example ACL implementation. (#12814)"
This reverts commit 381fc553fc9ab589b49149f8cf199d7923da17b8 because
that code was never reviewed and clearly wasn't tested either (it's
comparing TLV::Tag values to tag numbers, which can't possibly work).
