+++ src/access/BUILD.gn
@@ -20,7 +20,6 @@ static_library("access") {
   sources = [
     "AccessControl.cpp",
     "AccessControl.h",
-    "AccessControlEntryCodec.h",
     "AuthMode.h",
     "Privilege.h",
     "RequestPath.h",
